Fix masked labels filling for ambiguous cells (#97)
* Fix labels filling for ambiguous cells
Filling accidentally assigned a true value to cells that were never clustered. This properly assigns these cells a np.nan label

* Correct np.ma.filling and np.nan handling in plot.py.

+++ src/doubletdetection/doubletdetection.py
@@ -228,7 +228,7 @@ class BoostClassifier:
             with np.errstate(invalid='ignore'):  # Silence numpy warning about NaN comparison
                 self.voting_average_ = np.mean(np.ma.masked_invalid(self.all_p_values_) >= p_thresh,
                                                axis=0)
-                self.labels_ = np.ma.filled(self.voting_average_ >= voter_thresh, np.nan)
+                self.labels_ = np.ma.filled((self.voting_average_ >= voter_thresh).astype(float), np.nan)
                 self.voting_average_ = np.ma.filled(self.voting_average_, np.nan)
         else:
             # Find a cutoff score
